# tests/fixtures/studio_contracts/openapi/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ArtifactListResponse": {
        "properties": {
          "episode_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "ArtifactListResponse",
        "type": "object"
      },
...
      "DeriveRevisionRequest": {
        "additionalProperties": true,
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "episode_id": {
            "title": "Episode Id",
            "type": "string"
          },
          "expected_optimistic_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Optimistic Version"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "invalidation_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalidation Intent"
          },
          "new_content_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "New Content Hash",
            "type": "string"
          },
          "parent_revision_id": {
            "title": "Parent Revision Id",
            "type": "string"
          },
          "schema_version": {
            "default": "studio.command/v1",
            "title": "Schema Version",
            "type": "string"
          },
          "series_id": {
            "title": "Series Id",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "episode_id",
          "series_id",
          "parent_revision_id",
          "new_content_hash"
        ],
        "title": "DeriveRevisionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LockScreenplayRequest": {
        "additionalProperties": true,
        "properties": {
          "episode_id": {
            "title": "Episode Id",
            "type": "string"
          },
          "expected_content_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Expected Content Hash",
            "type": "string"
          },
          "expected_optimistic_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Optimistic Version"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "schema_version": {
            "default": "studio.command/v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "episode_id",
          "revision_id",
          "expected_content_hash"
        ],
        "title": "LockScreenplayRequest",
        "type": "object"
      },
...
      "RecordApprovalRequest": {
        "additionalProperties": true,
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "artifact_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Artifact Hash",
            "type": "string"
          },
          "checkpoint": {
            "minLength": 1,
            "title": "Checkpoint",
            "type": "string"
          },
          "decision": {
            "minLength": 1,
            "title": "Decision",
            "type": "string"
          },
          "episode_id": {
            "title": "Episode Id",
            "type": "string"
          },
          "expected_optimistic_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Optimistic Version"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "schema_version": {
            "default": "studio.command/v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "episode_id",
          "revision_id",
          "checkpoint",
          "artifact_hash",
          "decision"
        ],
        "title": "RecordApprovalRequest",
        "type": "object"
      },
...
      "SelectIdeaRequest": {
        "additionalProperties": true,
        "properties": {
          "candidate_id": {
            "minLength": 1,
            "title": "Candidate Id",
            "type": "string"
          },
          "episode_id": {
            "title": "Episode Id",
            "type": "string"
          },
          "expected_content_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Expected Content Hash",
            "type": "string"
          },
          "expected_optimistic_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Optimistic Version"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "schema_version": {
            "default": "studio.command/v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "episode_id",
          "revision_id",
          "candidate_id",
          "expected_content_hash"
        ],
        "title": "SelectIdeaRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Modular Monolith API V2 Production Application",
    "title": "WindAgent V2 API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/video-production/assets/{asset_id}/artifacts/{artifact_key}": {
      "get": {
        "description": "Get authorized access token / metadata for preview derivative delivery (UI27, UI28).",
        "operationId": "get_authorized_artifact_api_v2_video_production_assets__asset_id__artifacts__artifact_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Authorized Artifact Api V2 Video Production Assets  Asset Id  Artifacts  Artifact Key  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Authorized Artifact",
        "tags": [
          "production-assets"
        ]
      }
    },
...
    "/api/v3/studio/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_api_v3_studio_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Artifact Api V3 Studio Artifacts  Artifact Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact",
        "tags": [
          "studio-v3",
          "studio-artifacts"
        ]
      }
    },
...
    "/api/v3/studio/episodes/{episode_id}": {
      "get": {
        "operationId": "get_episode_api_v3_studio_episodes__episode_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Episode Api V3 Studio Episodes  Episode Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Episode",
        "tags": [
          "studio-v3",
          "studio-episodes"
        ]
      }
    },
    "/api/v3/studio/episodes/{episode_id}/approvals": {
      "post": {
        "operationId": "record_approval_api_v3_studio_episodes__episode_id__approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-WindAgent-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Windagent-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Record Approval Api V3 Studio Episodes  Episode Id  Approvals Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Approval",
        "tags": [
          "studio-v3",
          "studio-decisions"
        ]
      }
    },
    "/api/v3/studio/episodes/{episode_id}/artifacts": {
      "get": {
        "operationId": "list_artifacts_api_v3_studio_episodes__episode_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Artifacts",
        "tags": [
          "studio-v3",
          "studio-artifacts"
        ]
      }
    },
    "/api/v3/studio/episodes/{episode_id}/idea-selection": {
      "post": {
        "operationId": "select_idea_api_v3_studio_episodes__episode_id__idea_selection_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectIdeaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Select Idea Api V3 Studio Episodes  Episode Id  Idea Selection Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Idea",
        "tags": [
          "studio-v3",
          "studio-decisions"
        ]
      }
    },
    "/api/v3/studio/episodes/{episode_id}/revisions": {
      "post": {
        "operationId": "derive_revision_api_v3_studio_episodes__episode_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-WindAgent-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Windagent-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeriveRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Derive Revision Api V3 Studio Episodes  Episode Id  Revisions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Derive Revision",
        "tags": [
          "studio-v3",
          "studio-decisions"
        ]
      }
    },
...
    "/api/v3/studio/episodes/{episode_id}/screenplay-lock": {
      "post": {
        "operationId": "lock_screenplay_api_v3_studio_episodes__episode_id__screenplay_lock_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockScreenplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Lock Screenplay Api V3 Studio Episodes  Episode Id  Screenplay Lock Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lock Screenplay",
        "tags": [
          "studio-v3",
          "studio-decisions"
        ]
      }
    },
...
  }
}